move tests files
