+++ src/web/data/obstacle_avoidance_data/__tests__/exclusion-zone-set.test.ts
@@ -1,5 +1,5 @@
-import { ExclusionZone, ExclusionZoneSet } from "../exclusion-zone-set";
-import { GeographicCoordinate } from "../../../../types/protobuf-types";
+import { ExclusionZone, ExclusionZoneSet } from "../exclusion_zones/exclusion-zone-set";
+import { GeographicCoordinate } from "../../../types/protobuf-types";
 
 function coord(lat: number, lon: number): GeographicCoordinate {
     return { lat, lon };
